Move Strategy Toplevels to separate folder
